requirements.txt
dbus_python==1.2.18
pick==2.4.0
PyGObject==3.42.1
